requirements.txt
# ===============================================
# Airflow Providers (베이스 이미지에 Airflow 2.8.1 이미 설치됨)
# ===============================================
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-apache-kafka==1.3.1

# ===============================================
# Database
# ===============================================
psycopg2-binary==2.9.9

# ===============================================
# Data Processing
# ===============================================
pandas==2.1.4
numpy==1.26.3

# ===============================================
# HTTP & API
# ===============================================
requests==2.31.0
httpx==0.26.0

# ===============================================
# Async & Rate Limiting
# ===============================================
aiohttp==3.9.1
# asyncio는 Python 3.11 내장 모듈

# ===============================================
# Utilities
# ===============================================
python-dotenv==1.0.0
pyyaml==6.0.1
